Moved global_vas_and_funcs.
