+++ python/global_vals_funcs.py
